OM-56524: Hoisted `getEconomy` to `ActionImpl` to avoid code duplication
Every action "base" class as well as some leaf action classes defined an
economy field and a `getEconomy` method. At the early design stages not
all actions seemed to require this, but by now we have found they do, so
it's time to make it official and allow using this method without
downcasting.

That's not to say we will never refactor actions to make them economy-
independent. We might need to do that to avoid keeping the whole economy
until the next cycle. But keeping the duplication and downcasts in the
meantime makes maintenance more difficult and blocks other refactoring
efforts.

Also:

- Added some annotations.
- Removed the (unused) `type_` field from `ActionImpl`.
- Fixed some indentation.
- Fixed the name of a parameter in Javadoc to match the one in the
  method signature.