OM-55947: Removed `ReplayActions.translateShoppingList` and related tests
That method never translated shopping lists correctly as it depended on
quantities bought never changing between rounds. It has been replaced
by a local lambda in `replayActions` that depends on OIDs instead.

Some further investigation is necessary to ensure the shopping list OIDs
we currently use have the required properties for a correct translation.

Also removed `ReplayActions.translateCommoditySold` that was never used
outside tests. `translateTrader` will be removed in a future commit as
it's still being used in a few places.