OM-55947: Removed code exclusively used in preserving provision OIDs.
There was some code that attempted to preserve the OIDs of provisioned
traders between rounds. But that code created a new OID for the
provisioned traders each round and was just preserving the OID from the
time the ActionTOs were generated, until the next time they were
generated.

I asked Nitya who also discussed it briefly with Shravan. It seems we
don't need the provisioned traders to keep the same OID from one round
to the next.