OM-55947: Moved some responsibilities of `port` to an `isValid` method.
The first refactoring step aimed to replace the long if-elseif chain in
`replayActions` with dynamic dispatch. After moving each case to a
different method implementation, it became apparent that the kind of
checks `replayActions` was doing during porting of actions to a new
economy are self-contained and useful in their own right. They have been
moved to a separate `isValid` virtual method that we can also use in
tests to check actions returned by market algorithms and/or consult
before `take` like we do for `actionTaken`.

Also:

- Added some additional validity checks to `Activate` and `Deactivate`
  regarding current trader state.
- Removed an unused field that appears to be a remnant of a previews
  refactoring.