OM-55947: Replaced market field with basket in state-change actions.
When `Activate` and `Deactivate` classes were first written we thought
we might need to keep around the market that was analysed to create the
action. After reviewing the code a few years later I discovered we only
actually need the basket bought in that market. That's what's used
internally in analysis library and also what's being sent back as
`ActionTO`s.

By refactoring those classes and their corresponding shared base class
to remove the source market field and replace it with a triggering
basket field we avoid having to pass a `null` market in cases when the
seller-to-be-deactivated doesn't have any customers. That's because the
basket sold of that seller is appropriate in this case. Even if it
wasn't, an empty Basket was substituted by old code anyway, when market
was `null`. A special case of the above is when a trader used to be in
a market when the `Deactivate` action was created but is no longer when
the `Deactivate` action is being replayed.

The name of the new field was chosen to match that of the corresponding
one in `ActionTO`s and `ActionDTO`s.

From reviewing XL code using the getter in the corresponding
`ActionDTO`s it appears that there wasn't a significant special case for
the empty list of triggering commodities that needs to be preserved.
(There were some checks for it but the code ended up mapping it to the
empty string or a singleton list with trivial content in the end.)

Also:

- Changed all clients to pass the basket instead of the market to the
  `Activate` and `Deactivate` constructors.
- When the data flow was from the basket to the market I used the basket
  directly instead of getting it from the market.
- When a method parameter of type `Market` was only used to construct
  a state-change action, it was itself changed to `Basket`.
- Updated a few tests that didn't make sense now that `null` isn't a
  valid argument.
- Removed a redundant `return` statement.
- Removed a redundant nullability check from `Basket`.