OM-57366: Modified ReplayActions to keep the Topology, not just the map.
`ReplayActions` used to store a trader <-> OID `BiMap` from the old
topology for use in mapping traders of the old topology to matching ones
in the new during the replaying of actions of next analysis cycle.

Unfortunately this map isn't enough to port all types of actions. In
particular `Move` actions require mapping shopping lists as well to
destination topology.

In addition to that it's the `Topology` class that "knows" about OIDs
and is responsible for keeping those maps up-to-date and consistent with
the corresponding `Economy`.

To retain all the required data and avoid breaking `Topology`
encapsulation, but also to pave the way for future refactoring efforts
that will have to remove the 'topology' field from `Economy` that was
never supposed to know anything about topologies, I'm replacing the
`traderOids_` map with the full topology that contains it.

This also unblocks efforts to replace `ReplayActions.translateTrader`.

The old accessors for the 'traderOids' bimap didn't have javadoc nor
tests associated with them. I'll have to add those in a later commit.
Also, since topology and actions list have to match, it's advisable to
set them using the same method.

This is the `xl` part of the change.