OM-56791: Fix stitching operation comparisons for sorting
Background:

* The Java Comparator interface requires implementations to maintain 3 invariants:

1. The implementor must ensure that sgn(compare(x, y)) == -sgn(compare(y, x)) for all x and y. (This implies that compare(x, y) must throw an exception if and only if compare(y, x) throws an exception.)
2. The implementor must also ensure that the relation is transitive: ((compare(x, y)>0) && (compare(y, z)>0)) implies compare(x, z)>0.
3. Finally, the implementor must ensure that compare(x, y)==0 implies that sgn(compare(x, z))==sgn(compare(y, z)) for all z.

We appear to be violating invariant #3. For example:

StringToStringDataDrivenStitchingOperation[Cisco HyperFlex]<=>StringToStringDataDrivenStitchingOperation[GCP]=0
StringToStringDataDrivenStitchingOperation[GCP]<=>ListStringToListStringDataDrivenStitchingOperation[HPE 3PAR]=-1
StringToStringDataDrivenStitchingOperation[Cisco HyperFlex]<=>ListStringToListStringDataDrivenStitchingOperation[HPE 3PAR]=0

There are actually quite a few examples of this problem in the StandardProbeOrdering. It seems like this problem went unnoticed for a long time because Java’s default sort implementation (TimSort - https://en.wikipedia.org/wiki/Timsort) does not guarantee it will detect all violations to the invariants it will require. I ran an experiment to confirm this.

When the bug does occur, it entirely prevents broadcast so this is a very serious bug with no apparent fix so this is a very serious bug. In the best case we just stitch in the wrong order without knowing it which is also a fairly serious bug.

Solution:

We topologically sort the standard probe and target types to provide a total ordering of these. Because target types are only looked at in the ordering in the case when a probe type matches, this is sufficient to provide a correct ordering. Note that probes can register with unknown probe types and/or target types. These unknown types will always have their operations stitched last. If in the future we want to permit probes to provide more control of their operation stitching order when they register, we can adjust this behavior.

To make it easier to see the order in which main stitching operations (where operations are applied per-target) are applied, I have also updated logic in the stitching journal to include information about which specific probe type and target type are being addressed by a specific operation. For example:

-------------- START: StringToStringDataDrivenStitchingOperation ---------------
----------- Storage Browsing/vCenter Storage Browsing/73364095354352 -----------
----------------- vsphere-dc11.eng.vmturbo.com/73387802430112 ------------------

Testing Done:
* Built all of XL and ran tests for stitching and topology-processor.
* Updated related stitching and topology-processor unit tests.
* New tests in stitching and topology-processor for new behavior.
* Was able to reproduce the bug by loading diagnostics and getting the error "Comparison method violates its general contract!". After applying this fix, we are able to successfully broadcast.