OM-56441: XL Target: Update probe-dependent Target fields when probe registers with a new ProbeInfo.
Action capabilities are provided by the probe in its ProbeInfo, sent over as part of the registration protocol.
When we create a target we snapshot the ProbeInfo for that target and use it to populate any fields related to the ProbeInfo. When we look at the target - e.g. in the ProbeActionCapabilitiesApplicatorEditor - we take its cached ProbeInfo to determine how to set the movable/scalable/etc. flags on entities. The problem is that if the probe related to the target re-registered with an updated ProbeInfo we would not get it via the Target::probeInfo message.

Fix
===
Move the code dealing with ProbeInfo and probe-info derived properties in Target to a helper method, and call that helper method:
1) When creating the target, with the ProbeInfo of the probe at the time of creation.
2) Whenever a probe registers itself with a new ProbeInfo. This is done by making the CachingTargetStore a ProbeStoreListener, and forwarding the probe registration message to each target associated with the probe.

Reviewed At: https://rbcommons.com/s/VMTurbo/r/38176