OM-57367: Removed unused methods from `ReplayActions` and `Suspension`.
set/getRolledBackSuspensionCandidates of `ReplayActions` and
set/getRolledBack of `Suspension` were used to circle around an empty
set, but were otherwise unused. After consulting with Shravan and
reading https://rbcommons.com/s/VMTurbo/r/14235/, were the methods in
`ReplayActions` were introduced, I decided to remove them.

The review in particular, didn't mention why they were added but
appeared to be concerned with performance. A comment in `Suspension`
alludes to an effort to save time by not trying to suspend traders that
had failed to suspend in a previous cycle but that was either never
implemented or never worked.

For now I'm removing related accessors, fields and more importantly the
`translateRolledbackTraders` method to ease maintenance and refactoring.