OM-57370: Modified the signature of a method to improve type safety.
`tryReplayDeactivateActions` could only work if all actions it accepted
as an argument were `Deactivate` actions. Yet the corresponding check
was in its sole caller. Now the signature correctly reflects how the
method is intended to be used and any casts are close to the respective
checks.

This will also make some future planned refactoring easier.