OM-57372: Modified TopologyEntitiesHandlerTest to use a partial mock.
`TopologyEntitiesHandlerTest` was mocking `Analysis` in a way that
essentially said: `getReplayActions` always returns the same object
regardless of calls to `setReplayActions`. That happened to work so far
because the same object was always used as an argument to the setter,
but that's not the purpose of a setter and blocks refactoring efforts.

The test was changed to instead call the real accessor methods, but even
that wasn't enough as the mocking framework doesn't actually initialise
the object, so the return value of the first call to `getReplayActions`
had to also be hard-coded to the current initialiser for the field.

Also let IntelliJ IDEA replace an explicit use of an iterator with an
enhanced for loop.