OM-57374: Made ReplayActions a parameter of Ede.generateActions (Legacy)
It used to be a field of `Ede` and the change was made to better reflect
the fact that it was only meant to be used in `Ede.generateActions` and
not other `Ede` methods.

The change also makes it easier to understand that the `ReplayActions`
object isn't used again after the call to `Ede.generateActions`. The old
code made it look like the object was reused after the above method
modified it, but in reality all its contents were clobbered by new
values.

Also:

- Removed 1 out of the 5 overloaded versions of `Ede.generateActions`
  that was apparently never used.

This is the opsmgr part of the change.