OM-57113-Change the configuration for ical4j library so it does not try to get the latest timezone definition
Problem
=========
Disconnect the OpsMgr from the internet
Save the schedule. You should see the exception

Diagnosis
========
The ical4j library is used for our calendar based behavior which is mainly used for finding the next occurrence of a schedule. This library has its own timezone registry which is used for finding the next occurrence of a schedule. However, there is one behavior that this registry by default that is undesirable. This behavior is that it tries to get the latest definition of a timezone from internet (since timezone definitions can change over time) whenever a timezone is requested. The undesirable part is that the timezone retrieval fails if it cannot connect to internet and no timezone is returned. Our code falls back on default OpsMgr timezone when null is returned which can be wrong if the timezone is different from default OpsMgr timezone.

Solution
========
This can behavior can be fixed by setting update as false in ical4j properties file. This way the library uses existing timezone definitions and does not try to update.