OM-55947: Made Deactivate actions into a separate list in ReplayActions.
Currently we do some relatively expensive pre- and post-processing
before and after processing deactivate actions which makes it difficult
to process them between other actions while replaying. In fact current
code was populating the actions list exclusively with `Deactivate`
actions and delegating processing to a deactivates-exclusive method.

To make it clearer that `Deactivate` actions are processed separately,
but also to allow `Deactivate` actions to be replayed during the first
analysis sub-cycle while `ProvisionBySupply`, `ProvisionByDemand` and
`Activate` are replayed during the second analysis sub-cycle, a new list
of actions was added to `ReplayActions` exclusively for `Deactivate`
actions.

In addition `tryReplayDeactivateActions` was made `public` to initiate
processing of the new list and is no longer called by `replayActions`.

Also:
- Annotated a few parameters as `@NonNull`.
- Removed the `ledger` parameter from `replayActions`.

This is the xl part of the change.