OM-55816: Enabled replay of provision and activate actions.
This was done in an attempt to speed up real-time analysis and
especially its 2nd sub-cycle by reusing some of the work that was done
during a previous cycle.

Essentially a list of `ProvisionBySupply` and `Activate` actions
produced during a real-time analysis cycle is kept and re-applied during
the next analysis cycle. These actions don't have to be regenerated
during the next cycle and that saves time. Replay of `ProvisionByDemand`
actions isn't supported yet.

The new stage in `Ede.generateActions` was provisionally named
"provision replay".

Initial measurements with BofA topology indicate provision algorithm is
sped up by ~60%, while the whole analysis cycle is sped up by ~30%.

Also:

- Extended some comments.
- Added some annotations.
- Replaced some field references with the corresponding getters.
- Added support for subsequent actions in `replayActions`.
- Added some additional logging.

This is the opsmgr part of the change.