OM-55121 [RB-38431] Fix for Plan Summary Stats calls ignoring entity type filters
Background

A stat request from the API can contain many different filters. The most common one is name (i.e. commodity name), which was already supported for plan stats calls.

Problem

One filter that was not supported was relatedEntityType. Setting this on a stats request for a plan had no effect, leading to several UI bugs due to API calls including incorrect results in the response. It also led to issues distinguishing between commodities with the same name but different entity types--and forced us to put in special case logic to work around this.

Solution

Add a new entity_type column to the mkt_snapshots_stats table (in History component). Update the plan ingestion code in History component to populate this. On the query side, changes to both API component and History component were required.