[OM-57521, RB-38660] SLA Capacity removed from default policy settings for groups
**Intent**
Remove SLA Capacity removed from default policy settings of Business Applications and Database Servers.

**Background**
When selecting a group of BA's or DB's in XL, SLA Capacity is shown as a default constraint.
By the way, if we see the policies in policy page, there is no such parameter.

**Problem**
SLA Capacity should not be on default policy setting for groups.

**Solution**
SLA Capacity deleted from EntitySettingsSpec and PostStitchingOperationLibrary classes for all entities.
Since this parameter is not displayed on policies page and cannot be added, we don't need any migration changes.

**Unit Tests**
No Unit tests needed.