OM-57840: Changed suspension throttling used in 2nd sub-cycle to default
Now suspension throttling used in 2nd market sub-cycle is no longer
configurable and instead hard-coded to 'default'. The purpose of the 2nd
market sub-cycle (which is only present for real-time analysis) is to
generate provision and activate actions to be shown in the UI, without
the corresponding move or other actions. Since the acceptance criteria
no longer consider the effect of a provision/activate action to the rest
of a market, we heavily depend on the suspension algorithm being able to
negate those actions by generating suspension actions that will be
collapsed with them. That can only be done when 'default' throttling is
selected, i.e. when more than one suspension per cluster is permitted,
because more than one provision/activation may have been generated per
cluster.

An alternative was also considered: when marking all other traders in a
cluster as non-suspendible after suspending a trader in the cluster
(done when 'cluster' throttling is selected) we could skip clones. The
later was abandoned because it wouldn't work well later when we start
supporting collapse between provision/activate and suspend actions that
have different but equivalent targets.

e.g. we may have 4 identical PMs in a cluster A, B, C and D, decide to
clone A and B to A' and B' and then suspend C and D. Provision of A' and
B' could be collapsed with suspensions of C and D and the final cluster
would have 4 hosts. If we retain 'cluster' throttling and just skip
clones when marking co-sellers we would only generate suspension for C
and end up with a 5-host cluster.

Regarding performance, an alternative approach to improve performance,
compatible with 'default' throttling was found and is implemented and
described in next commit as it was deemed preferable.

This change was originally written by Shravan. I just wrote the commit
message and measured performance.