OM-56741: Copy container history from associated container spec RB:38583
The problem:
Cloud native containers and pods are ephemeral (short lifespan) and isomorphic (the only feature distinguishing Foo-1 from Foo-2 is its unique ID). When an individual container is restarted, behind the scenes the original container (say, "Foo-1234") and a new one is started in its place (say, "Foo-5678"). Furthermore, when sizing containers, all containers of the same type must be assigned the new size (similar to a consistent scaling group for VMs in the cloud). This leads to the problem that when a vertical scaling action is generated in Turbo, all containers of the same type (say, "Foo") are killed and replaced with new containers. This results in the loss of all history for all "Foo" instances. Since Market analysis is performed on historical utilization, this loss of history results in very poor recommendations after all containers are restarted. For example, since we forget historical peaks in the utilization history we may resize down far too aggressively if present utilization drops once all the containers are restarted.

The Solution:
* To resolve these problems (and others) that result from forgetting history for ephemeral entities, we store history for the ephemeral entities on durable entities and then transfer that shared history onto the individual entities at topology processor broadcast time.
* For additional details on the design for retaining history for ephemeral entities see: the bug page for OM_56130 which contains additional design details.
* Note that this class does no maintenance, settings lookups, etc. Instead, all the work for these functions is done on the entities where we aggregate shared ephemeral entity history (ie CONTAINER_SPEC). This stage performs a very simple and naive copy of that shared history onto the individual ephemeral entities that contribute to that shared history. Entities inserted or modified by plan scenario changes should set up their relationships so that they get the correct behavior from this stage.

Testing Done:
* New unit tests.
* Verified the new stage is executed during topology broadcast.