OM-57996: Only override container resizable when appropriate RB:38882
The Problem:
In testing changes for the new container model, we noticed an unexpectedly large reduction in the number of resize actions we were getting on containers vs with the old container model.

The Solution:
It turns out that with the new container model, we expected the probe to be able to send ContainerSpec commodities without setting the resizable flag on the commodities it was selling. Then the resizable flag might be set at a later point in the topology-processor if it had insufficient data to meet the minimum observation period. This value could then be transferred onto containers only if it was set.

However, we see that in practice the topology-processor blindly copies the resizable value from SDK DTO onto TopologyDTO even when it is not set, resulting in the transfer of the default value of false. As a result, we cannot rely on this value being unset later in the pipeline and we have to:

1. Have the probe set resizable=true on ContainerSpec commodities (done in a separate review in the kubeturbo codebase).
2. update the logic that copies this flag from ContainerSpec commodities onto Container commodities to only override the Container resizable value when the Container value is unset or true (this review).

Testing Done:
* Updated associated unit tests to check for condition.
* Yue tested in-app with the corresponding kubeturbo changes and we see a much more reasonable set of resize actions for the environment we were testing on.