OM-57840: Modified 2nd sub-cycle to only consider clones for suspension.
Currently we don't support collapsing provision/activate actions with
suspend actions that have a different, but equivalent, target. We also
discard any suspend actions that are generated in the 2nd market sub-
cycle but not collapsed. This change was done to improve performance by
not spending time to suspend traders for which the suspend action will
be discarded.

An additional benefit is that this change further reduces apparent
over-provisioning in the real-time analysis. That's because some clones
are now selected for suspension that would otherwise not be preferred
because other traders (non-clones) had lower ROI. Since the projected
topology from the 2nd sub-cycle was never send back and suspension
actions for the non-clones were discarded and didn't reduce the number
of provisions, the total number of traders in the end result
communicated to the rest of the system was higher.

Regarding performance the results are summarised in the following table.
The change to 'default' throttling had a 20-25% performance penalty and
reduced the number of provisions by 12-13%, while the change to consider
only clones for suspension in the 2nd sub-cycle almost brought back
performance to original levels and further reduced the number of
provisions by another 4-7%.

Measurements were done in combination with the corresponding commits in
opsmgr repository. They were based on `7.22.1-SNAPSHOT` build
`20200513225333000` loaded with 2020-01-12 BofA diagnostics. Debug and
quote cache were disabled in all cases.

       Version      | SNM | Execution Time | Provision Actions
--------------------|-----|----------------|------------------
7.22.1              | off |      1463s     |        1186
default throttling  | off |      1830s     |        1041
suspend only clones | off |      1665s     |         967
7.22.1              |  on |      3144s     |        1166
default throttling  |  on |      3795s     |        1014
suspend only clones |  on |      3093s     |         969