OM-56896- Do not return the accounts with no associated target  as members of dynamic group
Intent
======
Do not return the accounts with no associated target as members of group.

Problem
=======
A dynamic group of accounts based on the billing family indicates it has 4 members while it actually has 2 members.

Diagnosis
=========
We don't show the accounts with no targets added in the UI. However, we need to discover them. According to PM, we only need to show those accounts in the expanded targets page. However, we include them when getting the members of a dynamic group. This is wrong and accounts with no target should not be included as a group member.

Solution
========
We take a similar approach as 54869 where additional filter is added when making the call to repository component when the group is a dynamic group of accounts which prevents returning accounts with no target.