OM-56643,RB-39244- Populate the percentile value for plan entities if they have been cloned from another entity
Intent
======
If an entity is cloned from another entity use the percentile data use the data for the main entity to populate percentile from the cloned entity.

Problem
=======
If an entity is cloned from another entity, we need to use the original entity to lookup the percentile info for that entity. Right now, we don't do that.

Solution
========
I added ``getLiveTopologyFieldReference`` which returns the ``EntityCommodityFieldReference`` which point to the commodity in the live topology. This method is called when we need to get the key to lookup the value for the percentile from cache.