OM-57723-Check correct field when setting start time for UtilizationCountArray
Intent
======
Check start time rather than end time in the condition for setting the start time of UtilizationCountArray

Problem
=======
Hubert mentioned that he saw blobs with start time of zero in the database which should not happen

Diagnosis
=========
I was not able to reproduce the issue. However, Dmitry noticed part of a condition in the list of conditions for setting the start time is wrong and it looks at end time rather than start time when setting the start time.

Solution
========
Update the condition to look at start time rather than end time