OM-58808: Prevent resize up actions on container request commodities. RB:39332
**The problem**
We wish to allow resize down actions but prevent resize up actions on request (VCPURequest and VMemRequest) commodities on containers.

**Solution**
The market will not resize a commodity above its capacity upper bound. Capacity upper bound is set based on a commodity's max Threshold in the TopologyDTO. This change adds a stage to the topology pipeline to inject these threhsolds on the relevant container request commodities.

**Implementation**
* Rather than using a PostStitchingOperation, we create a new PipelineStage for this behavior because we may wish for different real-time vs plan behavior on request thresholds in the future.

**Testing Done**
* New unit tests
* Broadcast using a topology and saw relevant log message:
    [RequestCommodityThresholdsInjector]	: Injected 175 request commodity thresholds on 105 entities.
* Debugged in market component and saw that capacityUpperBound was being set appropriately for request commodities with the max threshold. We retain a capacity lower bound of 0 for now although we may wish to adjust this in the future.