[OM-58152] Stream responses to CloudCostStatRecord queries through grpc
RB: 39146

Background:
Previously the GetCloudCostStats RPC returned a single GetCloudCostStatsResponse. The GetCloudCostStatsResponse contains a list of CloudCostStatRecords, and each CloudCostStatRecord contains a list of inner StatRecords.

Problem:
For PWC, the API calls to populate the Storage Summary widget were failing. This was because there were so many volumes that the cloud cost data exceeded the maximum size for a GRPC message. Only a single CloudCostStatRecord was returned, but it contained an unmanageable number of inner StatRecords.

Solution:
+ Modify the GetCloudCostStats RPC to return a stream of GetCloudCostStatsResponse rather than just one.
+ Modify the implementation of the RPC call:
    + to partition the full list of StatRecords so that each CloudCostStatRecord doesn't contain more than maxNumberOfInnerStatRecords StatRecords
    + to partition the full list of CloudCostStatRecords based on the number of StatRecords in each, so that each GetCloudCostStatsResponse's CloudCostStatRecords don't contain more than maxNumberOfInnerStatRecords StatRecords in total.
+ Modify the callers of the RPC to iterate through each GetCloudCostStatsResponse to collect the CloudCostStatRecords

Note: I set the default value of maxNumberOfInnerStatRecords to be 43750 based on the largest observed RPC call from JnJ diags. Since the JnJ instance hasn't experienced this bug, we know that GRPC can handle at least that many.