[OM-58719] Delete Volume action CSV download: populate Region and Volume ID columns
RB:39531 Certain fields are required to populate certain columns of the Pending Actions CSV export. Previously, Delete Volume actions did not populate these fields, since they are not necessary in order to display the relevant information in the UI. With this change, the fields are populated to ensure completeness in the export.