OM-58808: Prevent resize up request actions in plan RB:39472
Problem:
Turbo should not generate any resize up actions, whether in real time in plans. The stage for applying thresholds to request commodities was added to the live pipeline but not the plan pipeline.

Solution:
Also apply the stage to the plan pipeline.

Testing Done:
Verified we no longer see resize up request actions on containers in plans.