[OM-59269] Delete Volume action CSV download: populate Resource column
RB:39576 The DeleteExplanation.size_kb field is required to populate the Resource column of the Pending Actions CSV export. Previously, Delete Volume actions did not populate this field, since it is not necessary in order to display the relevant information in the UI. With this change, the field is populated to ensure completeness in the export.