[OM-57719]: (Cost) Add CloudScopeStore and ComputeTierAllocationStore
Review: https://rbcommons.com/s/VMTurbo/r/39470/

**Background**
As part of CCA, we will begin to store per-entity allocation records, based on topology broadcasts. This commit adds the stores and tables necessary to store compute tier allocation records on a per entity basis.

**Cloud Scope Store**
This is the first introduction of a new CloudScopeStore and table. This table is meant to consolidate scoping information for cloud entities in a single table. It allows for indexing of all cloud scope attributes, while amortizing the cost of the index over all child tables. The expecation is that this table will be extended to entity costs and RI coverage in the future.