[OM-58174]: Introduce new connection_type: Controlls/ControlledBy
Typical examples are Kubernetes Controller (deployment, statefulset, etc.), AWS ASG, Azure ScaleSet, Node Pool, etc. In these cases, for example, the Kubernetes Deployment is the actual control point, while individual instance is the one participate in the market. The market action is generated on individual instance but the action execution should be on the deployment. The semantics of controls relationship is the controlled entities are the one participating in the market, but the actual action execution should be performed on the controller.