[OM-57829]: XL API: Add source target in the response when requesting for discovered groups
Background: Currently, when requesting for groups, there is no information about the source
target in the case the group is DISCOVERED. The information exists in the internal
representation, but is not added in the output GroupApiDTO.

Fix: Create a TargetApiDTO and add it to the output GroupApiDTO.

*Note*: There is a limitation currently in the api, which is that a GroupApiDTO can only
hold a single TargetApiDTO. But in the case of ResourceGroups there might be multiple
source targets. For now, as a convention if there are multiple targets always the one with
the lower oid will be returned (for consistency), and OM 59547 has been opened to properly
address this.

https://rbcommons.com/s/VMTurbo/r/39684/