[OM-58952]: Fix GroupTag Insertion RB-39801 Problem ============
While migrating diagnostics, we find an error that is caused by duplicate group tag insertion.

Solution
===========

Maintain a set that tracks the unique insertions. This is necessary as a check in the DB since the primary key comprises of the Group ID, tag key and tag value. Any time the same tag is attempted for re-insertion, log a debug message and skip over the record.