[OM-58360]: Initial batch of unit tests for extractor component
Review RB-39763: https://rbcommons.com/s/VMTurbo/r/39763

Several unit tests are provided for code that was initially committed
with none, to establish a code base for shared developmennt. Some
significant restructuring of the original code to improve testability,
and a handful of fixes for bugs revealed by the tests, are also
included.