[OM-59656]: Support decimal values for percentile value
Intent
======
Allow user to set aggressiveness percentile value to float value

Features Implemented
====================
Right now, the logic in topology processor converts the value set for percentile to integer. This is undesirable as we want to allow user to use percentile such as 99.5 or 99.9. Changed a few places in code so we support decimal values for percentile.