[OM-59639]: RB-39875 Remove the setting and placement policies for target that no longer discovered any policy
Intent
=====
- Remove the setting and placement policies for target that no longer discovered any policy

Problem
=======
In the BofA and PwC environment, we see an an error indicating that we have discovered policies with duplicate name which results in failure in group upload.

Diagnosis
=========
In code, we are missing the logic that removes the database entries for discovered policies that no longer discovered by any target. In those environment following scenario happened:
- The target A was there and discovered policy X
- The target A was removed.
- The entries for discovered policy A was still there but there are essentially ineffective as their associated groups have been deleted.
- The target A is added again and this time it has different oid (the changing in od typically don't happen but can happen occasionally because of change in the code or other reasons)
- We see the exception as we discover policy X again and this time associated to a target with different oid

Solution
========
I added a logic which removes the setting and placement policies associated to targets that no longer topology processor send them. This way we delete the policies associated to removed targets.