[OM-59877]: Fix cannot delete AD server (rb39895)
Problem:

After blanking ‘AD server’ in edit AD window of user management the txt string LDAP://:389
remains in the field and AD user can no longer log on using domain.

In ‘edit AD’ if a server is specified in the AD server field this specific AD server will be targeted rather than any other in the AD domain. If these details are saved and then re-edited to remove the AD server the string LDAP://:389 remains in the AD user field and it is not possible to log onto the instance using a created AD external user.

Investigation:

This is the HTTP body:

{"domainName":"corp.vmturbo.com","loginProviderURI":"","isSecure":false}

Backend is trying to create empty login provider(loginProviderURI), that’s why LDAP://:389.

Proposed fixes:

API component:  loginProviderURI is removed from ActiveDirectoryDTO if it’s empty.

Testing Done
Added unit tests
Integration test passed.