[OM-59747]: Fix the issue with autoscaling policies that they have same name
Intent
======
Ensuring the auto scaling and template exclusions policies discovered have unique name

Problem
=======
Group upload fails with error that there are duplicate policy names

Diagnosis
========
We don't allow policy with duplicate names. However, the customer had different autoscaling groups with same name across different regions. We create the policy using the name of that group and the target id. Therefore, that results in policies with duplicate names..

Solution
========
We can assume that the names for groups (not display name) are unique across targets. THerefore, I changed the topology process logic to create the name based on the group name, target id, and the policy type we are creating. This ensures that we never will face duplicates in policy names and if we do, there is a fault at probe level. Additionally, as we don't support policies name longer that 255, I added a logic which replaces last 55 character of a policy with hashed version of it if the length of policy name is bigger that 255.